+++ Projects/ClubberServices/source/Clubber.Backend.WebAPI/Controllers/EventController.cs
@@ -1,9 +1,10 @@
 ﻿using Clubber.Backend.Models.Model;
 using Clubber.Backend.Services.Logic.Services;
 using Clubber.Backend.WebAPI.Helpers;
-using MongoDB.Bson;
+using System;
 using System.Collections.Generic;
 using System.Configuration;
+using System.Linq;
 using System.Web.Http;
 
 namespace Eventber.WebAPI.Controllers
@@ -14,10 +15,10 @@ namespace Eventber.WebAPI.Controllers
 
         public EventController()
         {
-            //connection strings
+            // Connection strings
             string mongoConStr = ConfigurationManager.ConnectionStrings[Constants.MongoDB.MongoDBConectionString].ConnectionString;
             string redisConStr = ConfigurationManager.ConnectionStrings[Constants.RedisDB.RedisDBConectionString].ConnectionString;
-            //mongodb name
+            // MongoDB name
             string mongoDbName = ConfigurationManager.AppSettings[Constants.MongoDB.MongoDBDatabaseName];
 
             _iEventService = new EventService(mongoConStr, mongoDbName, redisConStr);
@@ -31,27 +32,58 @@ namespace Eventber.WebAPI.Controllers
         }
 
         // POST: api/Event
-        public void Post([FromBody]Event value)
+        public IHttpActionResult Post([FromBody]Event value)
         {
+            // Validation
             if (!ModelState.IsValid)
             {
-                return;
+                string messages = string.Join(Environment.NewLine, ModelState.Values
+                                           .SelectMany(x => x.Errors)
+                                           .Select(x => x.ErrorMessage));
+                return BadRequest(messages);
             }
 
+            // Call
             _iEventService.Add(value);
+
+            // Ret
+            return Ok();
         }
 
         // PUT: api/Event/5
-        public void Put(string id, [FromBody]Event value)
+        public IHttpActionResult Put(string id, [FromBody]Event value)
         {
+            // Validation
+            if (!ModelState.IsValid)
+            {
+                string messages = string.Join(Environment.NewLine, ModelState.Values
+                                        .SelectMany(x => x.Errors)
+                                        .Select(x => x.ErrorMessage));
+                return BadRequest(messages);
+            }
+
+            // Call
             value._id = id;
             _iEventService.Update(value);
+
+            // Ret
+            return Ok();
         }
 
         // DELETE: api/Event/5
-        public void Delete(string id)
+        public IHttpActionResult Delete(string id)
         {
+            // Validation
+            if (string.IsNullOrEmpty(id))
+            {
+                return BadRequest("Id is empty");
+            }
+
+            // Call
             _iEventService.Delete(id);
+
+            // Ret
+            return Ok();
         }
     }
 }
